Added player jump sounds, attack sounds, and created a coyote timer for smoother character movement.
Added an audio player that is an autoloader with the project and plays background music.

Created a main menu scene with a scrolling parallax background that is instanced within the main menu scene. The main menu scene includes a main menu options page and a load game page.

The settings_scene is instantiated within the main menu when the settings button is pushed.

Created a volume slider scene that is instantiated within the settings scene. Created two more audio buses that will handle music and sound effect volume. The Audio node from earlier will be modified so that these sound volumes can be controlled with the respective volume sliders.

Created a settings scene that can be instantiated within the main menu and will later be implemented in the game in the pause menu.